fix location of an area test
