+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.uni</groupId>
     <artifactId>unit</artifactId>
-    <version>0.9.0</version>
+    <version>0.9.1</version>
     <dependencies>
         <!-- test framework -->
         <dependency>
