+++ src/test/java/unit/area/Area2DTest.java
@@ -21,7 +21,6 @@
 
 package unit.area;
 
-import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
@@ -64,41 +63,6 @@ public class Area2DTest {
                     (width, height) -> {
                         MatcherAssert.assertThat(width, Matchers.is(resW));
                         MatcherAssert.assertThat(height, Matchers.is(resH));
-                    }
-                );
-                return null;
-            }
-        );
-    }
-
-    /**
-     * {@link Area2D#applyOn(Area, BiConsumer)} gives the right pos and size.
-     */
-    @Test
-    public void correctApply() {
-        // @checkstyle LocalFinalVariableName (4 lines)
-        final var resX = 42;
-        final var resY = 563;
-        final var resW = 34324;
-        final var resH = 233;
-        new Area2D(
-            new Pos2D(resX, resY),
-            new Size2D(resW, resH)
-        ).result(
-            (pos, size) -> {
-                pos.result(
-                    // @checkstyle ParameterName (1 line)
-                    (x, y) -> {
-                        MatcherAssert.assertThat(x, Matchers.is(resX));
-                        MatcherAssert.assertThat(y, Matchers.is(resY));
-                        return null;
-                    }
-                );
-                size.result(
-                    (width, height) -> {
-                        MatcherAssert.assertThat(width, Matchers.is(resW));
-                        MatcherAssert.assertThat(height, Matchers.is(resH));
-                        return null;
                     }
                 );
                 return null;
